@@ -9,14 +9,14 @@
     "logs": "firebase functions:log"
   },
   "engines": {
-    "node": "12"
+    "node": "16"
   },
   "main": "index.js",
   "dependencies": {
     "cors": "^2.8.5",
-    "firebase-admin": "^9.2.0",
-    "firebase-functions": "^3.11.0",
-    "nodemailer": "^6.6.1"
+    "firebase-admin": "^11.0.1",
+    "firebase-functions": "^3.22.0",
+    "nodemailer": "^6.7.8"
   },
   "devDependencies": {
     "firebase-functions-test": "^0.2.0"
