@@ -1,51 +1,53 @@
 const functions = require("firebase-functions");
-const admin = require("firebase-admin");
 const nodemailer = require("nodemailer");
-const cors = require("cors")({ origin: true });
-const creds = require("./config");
+const admin = require("firebase-admin");
 admin.initializeApp();
+const cors = require("cors")({ origin: true });
 
-var transport = {
-  host: "smtp.gmail.com",
-  port: 587,
+const gmailUser = functions.config().gmail.email;
+const gmailPass = functions.config().gmail.pass;
+
+var mailTransport = nodemailer.createTransport({
+  service: "gmail",
   auth: {
-    user: creds.USER,
-    pass: creds.PASS,
+    user: gmailUser,
+    pass: gmailPass,
   },
-};
-
-var transporter = nodemailer.createTransport(transport);
-
-transporter.verify((error, success) => {
-  if (error) {
-    console.log(error);
-  } else {
-    console.log("Server is ready to take messages");
-  }
 });
 
-exports.sendMail = functions.https.onRequest((request, response) => {
-  cors(req, res, () => {
-    var fname = req.query.fname;
-    var lname = req.query.lname;
-    var phone = req.query.phone;
-    var email = req.query.email;
-    var commentsQuestions = req.query.commentsQuestions;
-    var content = `First Name: ${fname}\nLast Name: ${lname}\nPhone: ${phone}\nE-mail: ${email}\nComment & Questions: ${commentsQuestions} `;
+exports.submit = functions.https.onRequest(async (req, res) => {
+  res.set("Access-Control-Allow-Origin", "*");
+  res.set("Access-Control-Allow-Methods", "GET, PUT, POST, OPTIONS");
+  res.set("Access-control-Allow-Headers", "*");
 
-    var mailOptions = {
-      from: fname + lname,
-      to: "benjiman.nolan1@gmail.com", // Change to email address that you want to receive messages on
-      subject: "New Contact Form",
-      text: content,
-    };
-  });
-
-  return transporter.sendMail(mailOptions, (err, data) => {
-    if (err) {
-      return res.send(err.toString());
-    } else {
-      return res.send("Sent");
-    }
-  });
+  if (req.method === "OPTIONS") {
+    res.end();
+  } else {
+    cors(req, res, async () => {
+      if (req.method !== "POST") {
+        return;
+      }
+      let fname = req.query.fname;
+      let lname = req.query.lname;
+      let phone = req.query.phone;
+      let email = req.query.email;
+      let commentsQuestions = req.query.commentsQuestions;
+      let content = `First Name: ${fname}\nLast Name: ${lname}\nPhone: ${phone}\nE-mail: ${email}\nComment & Questions: ${commentsQuestions} `;
+      const mailOptions = {
+        from: req.body.email,
+        replyTo: req.body.email,
+        to: gmailUser,
+        subject: `Website Contact Form: ${req.body.fname}`,
+        text: req.body.message,
+        html: `<p>${req.body.message}</p>`,
+      };
+      await mailTransport.sendMail(mailOptions).then(() => {
+        console.log(`New email sent to: ${gmailUser}`);
+        res.status(200).send({
+          isEmailSend: true,
+        });
+        return;
+      });
+    });
+  }
 });
